+++ breeze-unstable/src/main/kotlin/com/windea/breezeframework/core/domain/collections/WeakStack.kt
@@ -1,0 +1,3 @@
+package com.windea.breezeframework.core.domain.collections
+
+interface WeakStack